Вынес генерацию точки в отдельный презентер
