+++ src/main.js
@@ -1,18 +1,17 @@
 import HeaderTripInfo from './view/header-trip-info-view.js';
-import HeaderFilters from './view/header-time-filters-view.js';
+//mport HeaderFilters from './view/header-time-filters-view.js';
 
 import MockService from './service/mock-service.js';
 import DestinationsModel from './model/destinations-model.js';
 import OffersModel from './model/offers-model.js';
 import PointsModel from './model/points-model.js';
 
-import {generateFilter} from './mock/filter';
+//import {generateFilter} from './mock/filter';
 
 import {render, RenderPosition} from './framework/render.js';
 
 import EventsPresenter from './presenter/events-presenter.js';
 import FilterPresenter from './presenter/filter-presenter.js';
-
 
 const siteFiltersElement = document.querySelector('.trip-controls__filters');
 const siteTripMainElement = document.querySelector('.trip-main');
@@ -29,6 +28,7 @@ const eventsPresenter = new EventsPresenter({
   offersModel,
   pointsModel
 });
+
 
 const filterPresenter = new FilterPresenter({container: siteFiltersElement, pointsModel});
