+++ src/presenter/events-presenter.js
@@ -1,13 +1,11 @@
 import SortView from '../view/main-sort-view.js';
 import EventsList from '../view/main-events-list-view.js';
-
-import PointView from '../view/list-point-view.js';
-import EditPointView from '../view/list-edit-point-form.js';
 import EmptyList from '../view/list-empty.js';
 
 //import NewPointView from '../view/list-new-point-form-view.js';
 
-import { render, replace } from '../framework/render.js';
+import {render} from '../framework/render.js';
+import PointPresenter from './point-presenter.js';
 
 export default class EventsPresenter {
   #sortComponent = new SortView();
@@ -16,22 +14,27 @@ export default class EventsPresenter {
 
   #container = null;
 
+  #pointsModel = null;
   #destinationsModel = null;
   #offersModel = null;
-  #pointsModel = null;
-
   #points = [];
-
 
   constructor({container, destinationsModel, offersModel, pointsModel}) {
     this.#container = container;
+    this.#pointsModel = pointsModel;
     this.#destinationsModel = destinationsModel;
     this.#offersModel = offersModel;
-    this.#pointsModel = pointsModel;
-
     this.#points = [...this.#pointsModel.get()];
   }
 
+
+  #renderPoint(point) {
+    const pointPresenter = new PointPresenter({
+      destinationsModel: this.#destinationsModel,
+      offersModel: this.#offersModel,
+      eventList: this.#eventsListComponent});
+    pointPresenter.init(point);
+  }
 
   init() {
     if(this.#points.length === 0) {
@@ -42,58 +45,10 @@ export default class EventsPresenter {
     render(this.#sortComponent, this.#container);
     render(this.#eventsListComponent, this.#container);
 
+
     this.#points.forEach((point) => {
       this.#renderPoint(point);
     });
   }
 
-  #renderPoint = (point) => {
-    const pointComponent = new PointView({
-      point,
-      pointDestinations: this.#destinationsModel.getById(point.destination),
-      pointOffers: this.#offersModel.getByType(point.type),
-      onEditClick: pointEditClickHandler
-    });
-
-    const pointEditComponent = new EditPointView({
-      point,
-      pointDestinations: this.#destinationsModel.get(),
-      pointOffers: this.#offersModel.get(),
-      onResetClick: resetButtonClickHandler,
-      onSubmitClick: pointSubmitHandler
-    });
-
-    const replaceFormToPoint = () => {
-      replace(pointComponent, pointEditComponent);
-    };
-
-    const replacePointToForm = () => {
-      replace(pointEditComponent, pointComponent);
-    };
-
-    const escKeyDownHandler = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        evt.preventDefault();
-        replaceFormToPoint();
-        document.removeEventListener('keydown', escKeyDownHandler);
-      }
-    };
-
-    function pointEditClickHandler() {
-      replacePointToForm();
-      document.addEventListener('keydown', escKeyDownHandler);
-    }
-
-    function resetButtonClickHandler() {
-      replaceFormToPoint();
-      document.addEventListener('keydown', escKeyDownHandler);
-    }
-
-    function pointSubmitHandler() {
-      replaceFormToPoint();
-      document.removeEventListener('keydown', escKeyDownHandler);
-    }
-
-    render(pointComponent, this.#eventsListComponent.element);
-  };
 }
