+++ src/view/header-time-filters-view.js
@@ -1,8 +1,8 @@
 import AbstractView from '../framework/view/abstract-view.js';
 
-function createFilterTemplate(filter) {
-  const {type, isAvailability} = filter;
-  console.log(filter);
+function createFilterTemplate() {
+  //const {type, isAvailability} = filter;
+  //console.log(filter);
   return `<form class="trip-filters" action="#" method="get">
         <div class="trip-filters__filter">
           <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
